commit 2

+++ src/Practica/Opdracht_2.java
@@ -3,7 +3,7 @@ package Practica;
 public class Opdracht_2 {
 
     public static void main(String args[]) {
-        double i=1;
+        int i=1;
         while(i<=10) {
             System.out.println(i);
             i++;
